+++ CHANGES.md
@@ -45,6 +45,7 @@ Bugfix 🐛:
  - Fix emoji filtering not working
  - Fix issue of closing Realm in another thread (#725)
  - Attempt to properly cancel the crypto module when user signs out (#724)
+ - "ban" event are not rendered correctly (#716)
 
 Changes in RiotX 0.8.0 (2019-11-19)
 ===================================================